uploaded files mobile view
